feat: ui update
- Created a TextUI class to manage rendering text to the screen
- Updated positioning of UI and window size so that grid was centered
- Added high scores, lines cleared, and level statistics
- Stopped rendering text in Tetris.cpp and moved it to Game.cpp

+++ Tetris/Block.h
@@ -9,7 +9,7 @@ class Block
 {
 public:
 	Block();
-	void Draw(int offsetX, int offsetY);
+	void Draw(float offsetX, float offsetY);
 	void Move(int rows, int cols);
 	void Rotate(bool isClockwise);
 	std::vector<Position> GetCellPositions();
